bsp refactoring
