+++ emBODY/eBcode/arch-arm/board/amcbldc/bootloader/proj/amcbldc-bootloader.uvoptx
@@ -957,8 +957,8 @@
       <tvExp>0</tvExp>
       <tvExpOptDlg>0</tvExpOptDlg>
       <bDave2>0</bDave2>
-      <PathWithFileName>..\..\bsp\embot_hw_bsp_amcbldc.cpp</PathWithFileName>
-      <FilenameWithoutPath>embot_hw_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <PathWithFileName>..\..\bsp\embot_hw_led_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_led_bsp_amcbldc.cpp</FilenameWithoutPath>
       <RteFlg>0</RteFlg>
       <bShared>0</bShared>
     </File>
@@ -969,8 +969,8 @@
       <tvExp>0</tvExp>
       <tvExpOptDlg>0</tvExpOptDlg>
       <bDave2>0</bDave2>
-      <PathWithFileName>..\..\bsp\embot_hw_led_bsp_amcbldc.cpp</PathWithFileName>
-      <FilenameWithoutPath>embot_hw_led_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <PathWithFileName>..\..\bsp\embot_hw_gpio_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_gpio_bsp_amcbldc.cpp</FilenameWithoutPath>
       <RteFlg>0</RteFlg>
       <bShared>0</bShared>
     </File>
@@ -981,8 +981,44 @@
       <tvExp>0</tvExp>
       <tvExpOptDlg>0</tvExpOptDlg>
       <bDave2>0</bDave2>
-      <PathWithFileName>..\..\bsp\embot_hw_gpio_bsp_amcbldc.cpp</PathWithFileName>
-      <FilenameWithoutPath>embot_hw_gpio_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <PathWithFileName>..\..\bsp\embot_hw_button_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_button_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <RteFlg>0</RteFlg>
+      <bShared>0</bShared>
+    </File>
+    <File>
+      <GroupNumber>8</GroupNumber>
+      <FileNumber>41</FileNumber>
+      <FileType>8</FileType>
+      <tvExp>0</tvExp>
+      <tvExpOptDlg>0</tvExpOptDlg>
+      <bDave2>0</bDave2>
+      <PathWithFileName>..\..\bsp\embot_hw_can_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_can_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <RteFlg>0</RteFlg>
+      <bShared>0</bShared>
+    </File>
+    <File>
+      <GroupNumber>8</GroupNumber>
+      <FileNumber>42</FileNumber>
+      <FileType>8</FileType>
+      <tvExp>0</tvExp>
+      <tvExpOptDlg>0</tvExpOptDlg>
+      <bDave2>0</bDave2>
+      <PathWithFileName>..\..\bsp\embot_hw_flash_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_flash_bsp_amcbldc.cpp</FilenameWithoutPath>
+      <RteFlg>0</RteFlg>
+      <bShared>0</bShared>
+    </File>
+    <File>
+      <GroupNumber>8</GroupNumber>
+      <FileNumber>43</FileNumber>
+      <FileType>8</FileType>
+      <tvExp>0</tvExp>
+      <tvExpOptDlg>0</tvExpOptDlg>
+      <bDave2>0</bDave2>
+      <PathWithFileName>..\..\bsp\embot_hw_specialize_bsp_amcbldc.cpp</PathWithFileName>
+      <FilenameWithoutPath>embot_hw_specialize_bsp_amcbldc.cpp</FilenameWithoutPath>
       <RteFlg>0</RteFlg>
       <bShared>0</bShared>
     </File>
