+++ emBODY/eBcode/arch-arm/board/amcbldc/bootloader/proj/amcbldc-bootloader.uvprojx
@@ -675,11 +675,6 @@
           <GroupName>embot::hw::bsp</GroupName>
           <Files>
             <File>
-              <FileName>embot_hw_bsp_amcbldc.cpp</FileName>
-              <FileType>8</FileType>
-              <FilePath>..\..\bsp\embot_hw_bsp_amcbldc.cpp</FilePath>
-            </File>
-            <File>
               <FileName>embot_hw_led_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_led_bsp_amcbldc.cpp</FilePath>
@@ -688,6 +683,26 @@
               <FileName>embot_hw_gpio_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_gpio_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_button_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_button_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_can_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_can_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_flash_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_flash_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_specialize_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_specialize_bsp_amcbldc.cpp</FilePath>
             </File>
           </Files>
         </Group>
@@ -1434,11 +1449,6 @@
           <GroupName>embot::hw::bsp</GroupName>
           <Files>
             <File>
-              <FileName>embot_hw_bsp_amcbldc.cpp</FileName>
-              <FileType>8</FileType>
-              <FilePath>..\..\bsp\embot_hw_bsp_amcbldc.cpp</FilePath>
-            </File>
-            <File>
               <FileName>embot_hw_led_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_led_bsp_amcbldc.cpp</FilePath>
@@ -1447,6 +1457,26 @@
               <FileName>embot_hw_gpio_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_gpio_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_button_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_button_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_can_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_can_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_flash_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_flash_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_specialize_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>..\..\bsp\embot_hw_specialize_bsp_amcbldc.cpp</FilePath>
             </File>
           </Files>
         </Group>
