+++ emBODY/eBcode/arch-arm/board/amcbldc/bsp/embot_hw_application_bsp_amcbldc.cpp
@@ -40,9 +40,6 @@ using namespace embot::core::binary;
 
 #include "embot_hw_bsp_amcbldc_config.h"
 
-
-
-
 // --------------------------------------------------------------------------------------------------------------------
 // - board specific methods
 // --------------------------------------------------------------------------------------------------------------------
@@ -139,7 +136,6 @@ std::make_tuple(0x00470027, 0x484E500E, 0x20343356 ) // scheda jointlab camozzi
     }
 
 }
-
 
 
 // - end-of-file (leave a blank line after)----------------------------------------------------------------------------
