+++ emBODY/eBcode/arch-arm/board/amcbldc/bsp/embot_hw_button_bsp_amcbldc.cpp
@@ -1,0 +1,148 @@
+/*
+ * Copyright (C) 2025 iCub Tech - Istituto Italiano di Tecnologia
+ * Author:  Kevin Sangalli
+ * email:   kevin.sangalli@iit.it
+*/
+
+// --------------------------------------------------------------------------------------------------------------------
+// - public interface
+// --------------------------------------------------------------------------------------------------------------------
+
+#include "embot_hw_bsp.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+// - external dependencies
+// --------------------------------------------------------------------------------------------------------------------
+
+#include <cstring>
+#include <vector>
+#include <array>
+
+#include "embot_core_binary.h"
+#include "embot_core.h"
+#include "embot_hw_sys.h"
+
+#if defined(USE_STM32HAL)
+    #include "stm32hal.h"
+#else
+    #warning this implementation is only for stm32hal
+#endif
+
+
+using namespace std;
+using namespace embot::core::binary;
+
+// --------------------------------------------------------------------------------------------------------------------
+// - configuration of peripherals and chips. it is done board by board. it contains a check vs correct STM32HAL_BOARD_*
+// --------------------------------------------------------------------------------------------------------------------
+
+#include "embot_hw_bsp_amcbldc_config.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+// - support maps for the supported drivers
+// --------------------------------------------------------------------------------------------------------------------
+
+// - support map: begin of embot::hw::button
+
+#include "embot_hw_button.h"
+#include "embot_hw_button_bsp.h"
+#include "embot_hw_gpio_bsp.h"
+
+#if !defined(EMBOT_ENABLE_hw_button)
+
+namespace embot::hw::button {
+    
+    constexpr BSP thebsp { };
+    void BSP::init(embot::hw::BTN h) const {}
+    void BSP::onEXTI(const embot::hw::gpio::PROP &p) const {}
+    const BSP& getBSP() 
+    {
+        return thebsp;
+    }
+    
+}
+
+#else
+
+namespace embot::hw::button {
+       
+    // this button is attached directly to PB7 (SDA connector).
+    constexpr PROP btn1p = { .pressed = embot::hw::gpio::State::SET, .gpio = {embot::hw::GPIO::PORT::B, embot::hw::GPIO::PIN::seven}, 
+                             .pull = embot::hw::gpio::Pull::pulldown, .irqn = EXTI9_5_IRQn  };  
+
+    // this button is attached directly to PB15 (BREAK connector) but is available only for amcbldc rev b.
+    constexpr PROP btn2p = { .pressed = embot::hw::gpio::State::SET, .gpio = {embot::hw::GPIO::PORT::B, embot::hw::GPIO::PIN::fifteen}, 
+                             .pull = embot::hw::gpio::Pull::pulldown, .irqn = EXTI15_10_IRQn  };                               
+        
+    constexpr BSP thebsp {        
+        // maskofsupported
+        mask::pos2mask<uint32_t>(BTN::one) | mask::pos2mask<uint32_t>(BTN::two),        
+        // properties
+        {{
+            &btn1p, &btn2p          
+        }}        
+    };
+    
+    
+    void BSP::init(embot::hw::BTN h) const {}
+    
+    void BSP::onEXTI(const embot::hw::gpio::PROP &p) const
+    {
+        const embot::hw::GPIO gpio = embot::hw::gpio::getBSP().getGPIO(p);
+        switch(gpio.pin)
+        {
+            case embot::hw::GPIO::PIN::seven:
+            {
+                embot::hw::button::onexti(BTN::one);
+            } break;    
+
+            case embot::hw::GPIO::PIN::fifteen:
+            {
+                embot::hw::button::onexti(BTN::two);
+            } break;
+            
+            default:
+            {
+            } break;           
+        }              
+    }
+    
+    // we put in here the IRQHandlers + the exti callback
+
+    extern "C" {
+    
+        void EXTI9_5_IRQHandler(void)
+        {
+            HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_7);
+        }
+
+        void EXTI15_10_IRQHandler()
+        {
+            HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_15);      // the break
+            //HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_13);    // VCCOK_Pin
+            //HAL_GPIO_EXTI_IRQHandler(GPIO_PIN_14);    // VAUXOK_Pin
+        }
+        
+        void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin)
+        {
+             embot::hw::button::getBSP().onEXTI({nullptr, GPIO_Pin, nullptr});            
+        }        
+    }
+
+    const BSP& getBSP() 
+    {
+        return thebsp;
+    }
+    
+}
+
+#endif // button
+
+// - support map: end of embot::hw::button
+
+
+
+// - end-of-file (leave a blank line after)----------------------------------------------------------------------------
+
+    
+
