+++ emBODY/eBcode/arch-arm/board/amcbldc/bsp/embot_hw_flash_bsp_amcbldc.cpp
@@ -1,0 +1,120 @@
+/*
+ * Copyright (C) 2025 iCub Tech - Istituto Italiano di Tecnologia
+ * Author:  Kevin Sangalli
+ * email:   kevin.sangalli@iit.it
+*/
+
+// --------------------------------------------------------------------------------------------------------------------
+// - public interface
+// --------------------------------------------------------------------------------------------------------------------
+
+#include "embot_hw_bsp.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+// - external dependencies
+// --------------------------------------------------------------------------------------------------------------------
+
+#include <cstring>
+#include <vector>
+#include <array>
+
+#include "embot_core_binary.h"
+#include "embot_core.h"
+#include "embot_hw_sys.h"
+
+#if defined(USE_STM32HAL)
+    #include "stm32hal.h"
+#else
+    #warning this implementation is only for stm32hal
+#endif
+
+
+using namespace std;
+using namespace embot::core::binary;
+
+// --------------------------------------------------------------------------------------------------------------------
+// - configuration of peripherals and chips. it is done board by board. it contains a check vs correct STM32HAL_BOARD_*
+// --------------------------------------------------------------------------------------------------------------------
+
+#include "embot_hw_bsp_amcbldc_config.h"
+
+// --------------------------------------------------------------------------------------------------------------------
+// - support maps for the supported drivers
+// --------------------------------------------------------------------------------------------------------------------
+
+// - support map: begin of embot::hw::flash
+
+#include "embot_hw_flash_bsp.h"
+
+#if !defined(EMBOT_ENABLE_hw_flash)
+
+namespace embot::hw::flash {
+    
+    constexpr BSP thebsp { };
+    void BSP::init() const {}    
+    const BSP& getBSP() 
+    {
+        return thebsp;
+    }
+    
+}
+
+#else
+
+namespace embot::hw::flash::bsp {
+             
+    #if defined(EMBOT_ENABLE_hw_flash_DUALBANK)
+        #error EMBOT_ENABLE_hw_flash_DUALBANK should not be enabled
+        // must configure differently the bps, but ... dont define this mcro
+    #endif
+    // note:
+    // the stm32g4 has a page size of:
+    // - 4k when its flash is configured as single bank
+    // - 2k when its flash is configured as two banks
+    // we have used 4k because at the time did not have support for dual bank added only in nov 2022
+    
+    constexpr uint8_t numbanks {1};
+    constexpr size_t banksize {512*1024};
+    constexpr size_t pagesize {4*1024};
+    constexpr BankDescriptor bank01 { Bank::ID::one, 0x08000000, banksize, pagesize };
+    constexpr theBanks thebanks 
+    {
+        numbanks, 
+        { &bank01, nullptr }
+    }; 
+    
+    // on top of each other, with sizes:
+    constexpr std::array<uint32_t, 4> ss = {124*1024, 4*1024, 380*1024, 4*1024};
+    constexpr Partition btl {Partition::ID::bootloader,         &bank01,    bank01.address,         ss[0]}; 
+    constexpr Partition sha {Partition::ID::sharedstorage,      &bank01,    btl.address+btl.size,   ss[1]};
+    constexpr Partition app {Partition::ID::application,        &bank01,    sha.address+sha.size,   ss[2]}; 
+    constexpr Partition stg {Partition::ID::applicationstorage, &bank01,    app.address+app.size,   ss[3]}; 
+    
+    constexpr thePartitions thepartitions
+    {
+        { &btl, &sha, &app, &stg }
+    };
+        
+    constexpr BSP thebsp {        
+        thebanks,
+        thepartitions
+    };        
+    
+    void BSP::init() const {}
+    
+    const BSP& getBSP() 
+    {
+        return thebsp;
+    }        
+                     
+} // namespace embot::hw::flash::bsp {
+
+#endif // flash
+
+// - support map: end of embot::hw::flash
+
+
+// - end-of-file (leave a blank line after)----------------------------------------------------------------------------
+
+    
+
